requirements.txt
selenium
webdriver-manager
python-telegram-bot == 13.11
mysql-connector-python
pillow
opencv